2011-06-06  Dario Rodriguez  <dario.rodriguez@artica.es>
	* godmode/servers/manage_recontask.php: User with flag AW can manage recon
	task, only for their group.
	* godmode/servers/modificar_server.php: User with AW flag can view server
	status but only Pandora Admin can delete one of them.
	* godmode/servers/manage_recontask_form.php: User with flag AW can manage recon
	task, but only for their group.
	* godmode/menu.php: Allowing user with flag AW see manage server tab and 
	manage recon task subtab.



git-svn-id: https://svn.code.sf.net/p/pandora/code/trunk@4408 c3f86ba8-e40f-0410-aaad-9ba5e7f4b01f

+++ pandora_console/godmode/servers/modificar_server.php
@@ -20,7 +20,7 @@ require_once ("include/functions_servers.php");
 
 check_login();
 
-if (! check_acl ($config["id_user"], 0, "PM")) {
+if (! check_acl ($config["id_user"], 0, "AW")) {
 	db_pandora_audit("ACL Violation",
 		"Trying to access Server Management");
 	require ("general/noaccess.php");
@@ -105,7 +105,11 @@ else {
 		$table->head[3] = __('Type');
 		$table->head[4] = __('Started');
 		$table->head[5] = __('Updated');
-		$table->head[6] = __('Delete');
+		
+		//Only Pandora Administrator can delete servers
+		if (give_acl ($config["id_user"], 0, "PM")) {
+			$table->head[6] = __('Delete');
+		}
 		
 		foreach ($servers as $server) {
 			if ($server['status'] == 0) {
@@ -122,9 +126,13 @@ else {
 			$data[3] = $server['img'];
 			$data[4] = human_time_comparation ($server["laststart"]);
 			$data[5] = human_time_comparation ($server["keepalive"]);
-			$data[6] = '<a href="index.php?sec=gservers&sec2=godmode/servers/modificar_server&server_del='.$server["id_server"].'&amp;delete=1">';
-			$data[6] .= html_print_image ('images/cross.png', true, array ('title' => __('Delete'), 'onclick' => "if (! confirm ('" . __('Modules run by this server will stop working. Do you want to continue?') ."')) return false"));
-			$data[6] .= '</a>';
+
+			//Only Pandora Administrator can delete servers
+			if (give_acl ($config["id_user"], 0, "PM")) {
+				$data[6] = '<a href="index.php?sec=gservers&sec2=godmode/servers/modificar_server&server_del='.$server["id_server"].'&amp;delete=1">';
+				$data[6] .= html_print_image ('images/cross.png', true, array ('title' => __('Delete'), 'onclick' => "if (! confirm ('" . __('Modules run by this server will stop working. Do you want to continue?') ."')) return false"));
+				$data[6] .= '</a>';
+			}
 			
 			array_push ($table->data, $data);
 		}
